Fix a narrowing warning due to a type mismatch (size_t vs uint64).
lib/Transforms/Instrumentation/AddressSanitizer.cpp:1405:36: error: non-constant-expression cannot be narrowed from type 'uint64_t' (aka 'unsigned long long') to 'size_t' (aka 'unsigned int') in initializer list [-Wc++11-narrowing]
    getAllocaSizeInBytes(AI),
    ^~~~~~~~~~~~~~~~~~~~~~~~

llvm-svn: 196623

+++ llvm/include/llvm/Transforms/Utils/ASanStackFrameLayout.h
@@ -29,7 +29,7 @@ static const int kAsanStackRightRedzoneMagic = 0xf3;
 struct ASanStackVariableDescription {
   const char *Name;  // Name of the variable that will be displayed by asan
                      // if a stack-related bug is reported.
-  size_t Size;       // Size of the variable in bytes.
+  uint64_t Size;     // Size of the variable in bytes.
   size_t Alignment;  // Alignment of the variable (power of 2).
   AllocaInst *AI;    // The actual AllocaInst.
   size_t Offset;     // Offset from the beginning of the frame;
